produce_message/requirements.txt
kafka-python==2.0.0
